requirements.txt
Flask==3.0.2
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
